Refactor flexible agents and move functions to utils
