+++ core/flexible_agents.py
@@ -24,7 +24,6 @@ import yaml
 from dotenv import load_dotenv
 from google.adk.runners import Runner, types
 from google.adk.sessions import InMemorySessionService
-from jinja2 import Template
 
 # Add the project root to the Python path
 sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
@@ -36,322 +35,19 @@ from data_model import validate_configuration_file
 from utils import analyze_agent_structure, display_agent_readiness
 from utils.agent_utils import (
     collect_agent_execution_steps, display_execution_steps_summary, 
-    ExecutionStep, maintain_execution_status, report_finished_steps
+    ExecutionStep, maintain_execution_status, report_finished_steps,
+    save_results, display_results_summary
 )
+from utils.prompt_utils import synthesize_user_query_jinja2
+from utils.workflow_configuration import WorkflowConfiguration
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
 
 
-def load_job_config(job_config_path):
-    """Load job configuration from YAML or JSON file."""
-    with open(job_config_path, 'r') as f:
-        if str(job_config_path).endswith(('.yaml', '.yml')):
-            return yaml.safe_load(f)
-        else:
-            return json.load(f)
-
-
-def load_template_config(template_config_path):
-    """Load template configuration from YAML or JSON file."""
-    with open(template_config_path, 'r') as f:
-        if str(template_config_path).endswith(('.yaml', '.yml')):
-            return yaml.safe_load(f)
-        else:
-            return json.load(f)
-
-
-def read_input_file(file_path, input_type=None, **metadata):
-    """
-    Read a single input file and return its content and metadata.
-    
-    Args:
-        file_path (str or Path): Path to the input file
-        input_type (str): Explicit input type (overrides file extension)
-        **metadata: Additional metadata for the file
-        
-    Returns:
-        dict: Dictionary containing file_name, file_type, file_content, and metadata
-    """
-    file_path = Path(file_path)
-    
-    if not file_path.exists():
-        raise FileNotFoundError(f"Input file not found: {file_path}")
-    
-    document_reader = DocumentReader()
-    if document_reader.is_supported(file_path):
-        content = document_reader.read_document(file_path, as_markdown=True)
-    else:
-        with open(file_path, 'r', encoding='utf-8') as f:
-            content = f.read()
-
-    # Determine file type
-    if input_type:
-        file_type = input_type
-    else:
-        file_type = file_path.suffix.lstrip('.') if file_path.suffix else ''
-    
-    result = {
-        'file_name': file_path.name,
-        'file_type': file_type,
-        'file_content': content,
-        'file_size': len(content),
-        'full_path': str(file_path)
-    }
-    
-    # Add any additional metadata
-    result.update(metadata)
-    
-    return result
-
-
-def read_input_folder(folder_config, base_path=None):
-    """
-    Read all files from a folder (first level only) and return their content and metadata.
-    
-    Args:
-        folder_config (dict): Folder configuration with input_path and optional target_agent
-        base_path (Path): Base path for resolving relative paths
-        
-    Returns:
-        list: List of file data dictionaries from the folder
-    """
-    folder_path = Path(folder_config['input_path'])
-    if base_path and not folder_path.is_absolute():
-        folder_path = base_path / folder_path
-    
-    if not folder_path.exists():
-        raise FileNotFoundError(f"Input folder not found: {folder_path}")
-    
-    if not folder_path.is_dir():
-        raise ValueError(f"Input path is not a directory: {folder_path}")
-    
-    files_data = []
-    
-    # Read only first-level files (no subdirectories)
-    for file_path in folder_path.iterdir():
-        if not file_path.is_file():
-            continue
-            
-        # Skip hidden files
-        if file_path.name.startswith('.'):
-            continue
-            
-        try:
-            file_data = read_input_file(file_path, folder_config.get('input_type'))
-            # Add folder context to metadata
-            file_data['source_folder'] = str(folder_path)
-            files_data.append(file_data)
-            logging.info(f"📁 Read file from folder: {file_data['file_name']} ({file_data['file_size']} chars)")
-        except (UnicodeDecodeError, PermissionError) as e:
-            logging.warning(f"⚠️  Skipping file {file_path}: {e}")
-        except Exception as e:
-            logging.error(f"❌ Error reading file {file_path}: {e}")
-    
-    if not files_data:
-        logging.warning(f"⚠️  No files found in folder {folder_path}")
-    else:
-        logging.info(f"📂 Successfully read {len(files_data)} files from folder: {folder_path}")
-        
-    return files_data
-
-
-def append_content_to_agent_config(agent_config, target_agent_name, grouped_files):
-    """
-    Recursively find and append content to the target agent's instruction in the config.
-    
-    Args:
-        agent_config: The agent configuration dictionary
-        target_agent_name: Name of the target agent to append content to
-        grouped_files: List of file data dictionaries for this target agent
-        
-    Returns:
-        bool: True if target agent was found and content appended, False otherwise
-    """
-    def _append_to_agent_config(current_config):
-        if current_config.get('name') == target_agent_name:
-            # Found the target agent, append content to its instruction
-            current_instruction = current_config.get('instruction', '') or ''
-            
-            # Create grouped content with prefix
-            file_names = [f['file_name'] for f in grouped_files]
-            additional_content = f"\n\nFocus on the content from the following files: {', '.join(file_names)}\n"
-            
-            for file_data in grouped_files:
-                additional_content += f"\n--- Content from {file_data['file_name']} ---\n"
-                
-                # Use a more aggressive approach: replace all curly braces with HTML entities
-                # This completely prevents any template interpretation
-                escaped_content = file_data['file_content']
-                escaped_content = escaped_content.replace('{', '&#123;').replace('}', '&#125;')
-                escaped_content = escaped_content.replace('[', '&#91;').replace(']', '&#93;')
-                
-                # Wrap in code block with explicit language tag
-                additional_content += f"\n```python\n{escaped_content}\n```\n"
-                additional_content += f"\n--- End of {file_data['file_name']} ---\n"
-            
-            current_config['instruction'] = current_instruction + additional_content
-            logging.info(f"📎 Appended content from {len(grouped_files)} file(s) to agent config '{target_agent_name}'")
-            return True
-        
-        # If this is a composite agent, search sub_agents
-        if 'sub_agents' in current_config and current_config['sub_agents']:
-            for sub_agent_config in current_config['sub_agents']:
-                if _append_to_agent_config(sub_agent_config):
-                    return True
-        
-        return False
-    
-    return _append_to_agent_config(agent_config)
-
-
-def synthesize_user_query_jinja2(template_config, file_names, file_types, file_contents):
-    """
-    Synthesize user query using Jinja2 template.
-    
-    Args:
-        template_config: Template configuration dictionary
-        file_names: List of file names (strings)
-        file_types: List of file types (strings)  
-        file_contents: List of file contents (strings)
-    """
-    
-    # Get template content
-    template_content = template_config.get('template_content', '')
-    variable_mapping = template_config.get('template_variables', {})
-    
-    # Template variables
-    template_vars = {
-        k: v['default'] for k, v in variable_mapping.items()
-    }
-
-    logging.debug(f"Using template variables: {template_vars}")
-
-    # Update template variables with file data arrays
-    template_vars.update({
-        'file_name': file_names,
-        'file_type': file_types, 
-        'file_content': file_contents
-    })
-
-    # Render template
-    template = Template(template_content)
-    rendered_query = template.render(**template_vars)
-
-    logging.debug(f"Rendered query length: {len(rendered_query)} characters")
-
-    return rendered_query
-
-
 emojis = ["👤", "🤖", "💡", "🔍", "⚙️", "📊", "🛠️", "📈", "📝", "✅", "🌟", 
           "🚀", "🎯", "🧩", "🔧", "📅", "💻", "🖥️", "📱", "🖨️", "🗂️", "🔒", 
           "🔑", "🧰", "🧪", "🧬", "🧫", "🧫", "🧪", "🧬"]
-
-
-def create_agent_from_config_file(config_path):
-    """
-    Complete workflow to create an agent from a configuration file.
-    
-    Args:
-        config_path (str or Path): Path to the configuration file
-        
-    Returns:
-        The instantiated agent or None if creation failed
-    """
-    config_path = Path(config_path)
-    
-    logging.info(f"Initializing agent: {config_path.name}")
-
-    # Validate configuration
-    try:
-        _, warnings = validate_configuration_file(config_path)
-        if warnings:
-            logging.warning(f"Configuration has {len(warnings)} warnings")
-    except Exception as e:
-        logging.error(f"Configuration validation failed: {e}")
-        return None
-    
-    # Create agent
-    try:
-        agent = create_agent_from_config(str(config_path))
-        logging.info(f"Agent '{agent.name}' created successfully")
-    except Exception as e:
-        logging.error(f"Agent creation failed: {e}")
-        return None
-    
-    # Analyze structure and display readiness
-    analyze_agent_structure(agent)
-    display_agent_readiness(agent)
-    
-    return agent
-
-
-def _save_results(input_files_data, agent, event_count, final_responses: Dict[str, str], job_config: dict):
-    """Helper function to save agent execution results to files."""
-    output_config = job_config.get('output_config', {})
-    output_dir = Path(__file__).parent.parent / output_config.get('output_directory', 'output')
-    output_dir.mkdir(exist_ok=True)
-    
-    timestamp_format = output_config.get('timestamp_format', '%Y%m%d_%H%M%S')
-    timestamp = datetime.datetime.now().strftime(timestamp_format)
-    current_time = datetime.datetime.now()
-    
-    # Generate filename based on input files
-    if len(input_files_data) == 1:
-        input_filename = Path(input_files_data[0]['full_path']).stem
-    else:
-        input_filename = f"multi_file_execution_{len(input_files_data)}_files"
-    
-    # Save text report
-    file_naming = output_config.get('file_naming', 'agent_execution_{input_filename}_{timestamp}')
-    output_file = output_dir / f"{file_naming.format(input_filename=input_filename, timestamp=timestamp)}.txt"
-    with open(output_file, 'w') as f:
-        f.write(f"Agent Execution Report\n")
-        f.write(f"=====================\n\n")
-        f.write(f"Execution Date: {current_time.strftime('%Y-%m-%d %H:%M:%S')}\n")
-        f.write(f"Input Files ({len(input_files_data)}):\n")
-        for i, file_data in enumerate(input_files_data, 1):
-            f.write(f"  {i}. {file_data['full_path']} ({file_data['file_size']} chars)\n")
-        f.write(f"Agent: {agent.name}\n")
-        f.write(f"Events Generated: {event_count}\n\n")
-        
-        # Include final responses
-        if final_responses:
-            f.write("\n\nFinal Responses:\n")
-            f.write("-" * 40 + "\n")
-            for author, response in final_responses.items():
-                f.write(f"{author}:\n")
-                f.write(response + "\n\n")
-    
-    # Save JSON report
-    json_output = {
-        "metadata": {
-            "execution_date": current_time.isoformat(),
-            "input_files": [
-                {
-                    "file_path": file_data['full_path'],
-                    "file_name": file_data['file_name'],
-                    "file_type": file_data['file_type'],
-                    "file_size": file_data['file_size'],
-                }
-                for file_data in input_files_data
-            ],
-            "agent_name": agent.name,
-            "events_generated": event_count,
-            "total_file_size": sum(f['file_size'] for f in input_files_data),
-        },
-        "execution_results": final_responses,
-        "content_analyzed": [file_data['file_content'] for file_data in input_files_data]
-    }
-    
-    json_file = output_dir / f"{file_naming.format(input_filename=input_filename, timestamp=timestamp)}.json"
-    with open(json_file, 'w') as f:
-        json.dump(json_output, f, indent=2)
-    
-    logging.info(f"📁 Output saved to: {output_file}")
-    logging.info(f"📄 JSON output saved to: {json_file}")
-
-    return str(output_file), str(json_file)
 
 
 async def run_agent(agent, user_query, job_config: dict):
@@ -415,7 +111,7 @@ def get_error_code_from_event(event):
     return error_code
 
 
-async def run_job(agent, input_file_paths, execution_steps: Dict[str, ExecutionStep], job_config: dict):
+async def run_job(agent, input_file_paths, execution_steps: Dict[str, ExecutionStep], workflow_config: WorkflowConfiguration):
     """
     Run the agent with input files for processing.
     
@@ -423,7 +119,7 @@ async def run_job(agent, input_file_paths, execution_steps: Dict[str, ExecutionS
         agent: The instantiated agent
         input_file_paths: List of paths to the files to process, or single path as string
         execution_steps: Dictionary of execution steps to track
-        job_config: Job configuration dictionary
+        workflow_config: WorkflowConfiguration instance for all configuration needs
         
     Returns:
         dict: Execution results with file paths and metadata
@@ -456,7 +152,7 @@ async def run_job(agent, input_file_paths, execution_steps: Dict[str, ExecutionS
             input_type = file_info.get('input_type')
             target_agent = file_info.get('target_agent')
             
-            file_data = read_input_file(file_path, input_type)
+            file_data = workflow_config.read_input_file(Path(file_path), input_type)
             input_files_data.append(file_data)
             
             if target_agent:
@@ -483,7 +179,7 @@ async def run_job(agent, input_file_paths, execution_steps: Dict[str, ExecutionS
     logging.info(f"Total size: {total_size} characters")
 
     # Create analysis request using Jinja2 template synthesis
-    analysis_config = job_config.get('analysis_config', {})
+    analysis_config = workflow_config.job_config.get('analysis_config', {})
     
     # Load template config - either from content or from file path
     if 'template_config_content' in analysis_config:
@@ -495,7 +191,7 @@ async def run_job(agent, input_file_paths, execution_steps: Dict[str, ExecutionS
         template_config_path = analysis_config.get('template_config_path')
         logging.info(f"Using template config path from job config: {template_config_path}")
         template_full_path = Path(__file__).parent.parent / template_config_path
-        template_config = load_template_config(template_full_path)
+        template_config = workflow_config.load_template_config(template_full_path)
     
     # Create analysis request using Jinja2 template synthesis (only for non-targeted files)
     if file_names:  # Only synthesize if there are non-targeted files
@@ -509,7 +205,7 @@ async def run_job(agent, input_file_paths, execution_steps: Dict[str, ExecutionS
         logging.info("All files are targeted to specific agents, using basic query")
 
     try:
-        response_generator, session = await run_agent(agent, user_query, job_config)
+        response_generator, session = await run_agent(agent, user_query, workflow_config.job_config)
         final_responses = {}
         event_count = 0
         
@@ -559,8 +255,8 @@ async def run_job(agent, input_file_paths, execution_steps: Dict[str, ExecutionS
         logging.info(f"Execution completed: {event_count} events generated")
 
         # Save results
-        output_file, json_file = _save_results(
-            input_files_data, agent, event_count, final_responses, job_config
+        output_file, json_file = save_results(
+            input_files_data, agent, event_count, final_responses, workflow_config.job_config
         )
         
         return {
@@ -578,24 +274,6 @@ async def run_job(agent, input_file_paths, execution_steps: Dict[str, ExecutionS
         return None
 
 
-def _display_results_summary(results):
-    """Helper function to display execution results summary."""
-    logging.info("\n" + "=" * 60)
-    logging.info("Execution Results Summary")
-    logging.info("=" * 60)
-
-    if isinstance(results, dict):
-        logging.info(f"Status: {results.get('status', 'unknown')}")
-        logging.info(f"Events Generated: {results.get('events_generated', 0)}")
-        logging.info(f"Response Length: {results.get('response_length', 0)} characters")
-        if results.get('output_file'):
-            logging.info(f"Text Output: {results['output_file']}")
-        if results.get('json_file'):
-            logging.info(f"JSON Output: {results['json_file']}")
-    else:
-        logging.warning("No results returned from agent execution")
-
-
 async def main_async_with_config(job_config_content: str, agent_config_content: str, template_config_content: str):
     """
     Main async function that creates and runs the flexible agent using YAML content directly.
@@ -610,106 +288,33 @@ async def main_async_with_config(job_config_content: str, agent_config_content: 
     """
     
     try:
-        # Parse job configuration from content
-        job_config = yaml.safe_load(job_config_content)
-        logging.info(f"Loaded job config: {job_config.get('job_name', 'Unknown')}")
+        # Initialize WorkflowConfiguration
+        workflow_config = WorkflowConfiguration()
         
-        job_name = job_config.get('job_name', 'Flexible Agent')
+        # Load configurations
+        workflow_config.load_job_config_from_content(job_config_content)
+        workflow_config.load_agent_config_from_content(agent_config_content)
+        workflow_config.load_template_config_from_content(template_config_content)
+        
+        job_name = workflow_config.job_config.get('job_name', 'Flexible Agent')
+        logging.info(f"Loaded job config: {workflow_config.job_config.get('job_name', 'Unknown')}")
         logging.info(f"{job_name} - Execution")
-
-        # Parse agent configuration from content
-        agent_config = yaml.safe_load(agent_config_content)
         
-        # Process input files and folders and group by target agent BEFORE creating agents
-        input_config = job_config.get('input_config', {})
-        
-        if 'input_files' not in input_config and 'input_folders' not in input_config:
-            logging.error("No input_files or input_folders found in input_config. Please configure at least one input source.")
+        # Process input files and folders using WorkflowConfiguration
+        try:
+            input_files, _ = workflow_config.process_input_files_and_folders()
+        except (ValueError, FileNotFoundError) as e:
+            logging.error(str(e))
             return 1
-            
-        input_files = []
-        targeted_files_by_agent = {}  # Group files by target agent
         
-        # Process input_files if present
-        if 'input_files' in input_config:
-            input_files_config = input_config['input_files']
-            logging.info(f"Processing {len(input_files_config)} input files:")
-            
-            for i, file_config in enumerate(input_files_config, 1):
-                file_path = Path(__file__).parent.parent / file_config['input_path']
-                target_agent = file_config.get('target_agent')
-                
-                input_files.append({
-                    'path': file_path,
-                    'input_type': file_config.get('input_type'),
-                    'target_agent': target_agent
-                })
-                
-                # If this file targets a specific agent, read it and group it
-                if target_agent:
-                    try:
-                        file_data = read_input_file(file_path, file_config.get('input_type'))
-                        if target_agent not in targeted_files_by_agent:
-                            targeted_files_by_agent[target_agent] = []
-                        targeted_files_by_agent[target_agent].append(file_data)
-                        target_info = f" -> {target_agent}"
-                    except FileNotFoundError as e:
-                        logging.error(str(e))
-                        return 1
-                else:
-                    target_info = ""
-                
-                logging.info(f"  {i}. {file_path} ({file_config.get('input_type', 'auto')}){target_info}")
-        
-        # Process input_folders if present
-        if 'input_folders' in input_config:
-            input_folders_config = input_config['input_folders']
-            logging.info(f"Processing {len(input_folders_config)} input folders:")
-            
-            for i, folder_config in enumerate(input_folders_config, 1):
-                folder_path = Path(__file__).parent.parent / folder_config['input_path']
-                target_agent = folder_config.get('target_agent')
-                
-                try:
-                    # Read all files from the folder
-                    folder_files_data = read_input_folder(folder_config, Path(__file__).parent.parent)
-                    
-                    # Add each file to the input_files list and process targeting
-                    for file_data in folder_files_data:
-                        input_files.append({
-                            'path': file_data['full_path'],
-                            'input_type': file_data['file_type'],
-                            'target_agent': target_agent
-                        })
-                        
-                        # If this folder targets a specific agent, group all its files
-                        if target_agent:
-                            if target_agent not in targeted_files_by_agent:
-                                targeted_files_by_agent[target_agent] = []
-                            targeted_files_by_agent[target_agent].append(file_data)
-                    
-                    target_info = f" -> {target_agent}" if target_agent else ""
-                    logging.info(f"  {i}. {folder_path} ({len(folder_files_data)} files){target_info}")
-                    
-                except (FileNotFoundError, ValueError) as e:
-                    logging.error(str(e))
-                    return 1
-        
-        # Modify agent configuration to include targeted file content
-        if targeted_files_by_agent:
-            logging.info(f"Modifying agent configuration for {len(targeted_files_by_agent)} targeted agent(s)...")
-            for target_agent, files in targeted_files_by_agent.items():
-                success = append_content_to_agent_config(agent_config, target_agent, files)
-                if not success:
-                    logging.warning(f"⚠️  Could not find target agent '{target_agent}' in configuration")
-        
-        logging.info("Creating agent from provided configuration...")
+        # Apply targeted file content to agent configuration
+        workflow_config.apply_targeted_files_to_agent_config()
         
         # Create agent from the modified configuration
-        agent = _create_agent_from_dict(agent_config)
+        agent = _create_agent_from_dict(workflow_config.agent_config)
         
         if agent is None:
-            logging.error("Failed to create agent. Exiting.")
+            logging.error("Failed to create agent")
             return 1
         
         # Analyze structure and display readiness
@@ -717,33 +322,20 @@ async def main_async_with_config(job_config_content: str, agent_config_content: 
         display_agent_readiness(agent)
         
         # Collect agent execution steps
-        execution_config = job_config.get('execution_config', {})
+        execution_config = workflow_config.get_execution_config()
         if execution_config.get('track_execution_steps', True):
-            logging.info(f"\nCollecting agent execution steps...")
             execution_steps = collect_agent_execution_steps(agent)
-            logging.info(f"💡 Total execution steps collected: {len(execution_steps)}")
             if execution_config.get('display_progress', True):
                 display_execution_steps_summary(execution_steps)
         else:
             execution_steps = {}
         
-        # Parse template configuration and modify job_config to include it
-        template_config = yaml.safe_load(template_config_content)
-        
-        # Temporarily modify job_config to include template content instead of path
-        original_analysis_config = job_config.get('analysis_config', {}).copy()
-        job_config['analysis_config'] = job_config.get('analysis_config', {}).copy()
-        job_config['analysis_config']['template_config_content'] = template_config
-        
-        results = await run_job(agent, input_files, execution_steps, job_config)
-        
-        # Restore original analysis_config
-        job_config['analysis_config'] = original_analysis_config
+        results = await run_job(agent, input_files, execution_steps, workflow_config)
         
         # Display results
-        report_config = job_config.get('report_config', {})
+        report_config = workflow_config.get_report_config()
         if results and report_config.get('display_results_summary', True):
-            _display_results_summary(results)
+            display_results_summary(results)
             if execution_config.get('track_execution_steps', True):
                 report_finished_steps(execution_steps)
         
@@ -776,12 +368,15 @@ async def main_async(job_name: str = "simple_code_improvement"):
         with open(job_config_path, 'r') as f:
             job_config_content = f.read()
         
-        job_config = yaml.safe_load(job_config_content)
-        logging.info(f"Loaded job config: {job_config.get('job_name', 'Unknown')}")
+        # Initialize WorkflowConfiguration
+        workflow_config = WorkflowConfiguration()
+        workflow_config.load_job_config_from_content(job_config_content)
+        
+        logging.info(f"Loaded job config: {workflow_config.job_config.get('job_name', 'Unknown')}")
 
         # Load agent configuration content
-        agent_config = job_config.get('agent_config', {})
-        config_path = Path(__file__).parent.parent / agent_config.get('config_path', 'config/agent/json_examples/simple_code_improvement.json')
+        agent_config_info = workflow_config.job_config.get('agent_config', {})
+        config_path = Path(__file__).parent.parent / agent_config_info.get('config_path', 'config/agent/json_examples/simple_code_improvement.json')
         logging.info(f"Loading agent config from: {config_path}")
         
         with open(config_path, 'r') as f:
@@ -793,7 +388,7 @@ async def main_async(job_name: str = "simple_code_improvement"):
                 agent_config_content = yaml.dump(agent_config_dict)
         
         # Load template configuration content
-        analysis_config = job_config.get('analysis_config', {})
+        analysis_config = workflow_config.job_config.get('analysis_config', {})
         template_config_path = analysis_config.get('template_config_path')
         template_full_path = Path(__file__).parent.parent / template_config_path
         logging.info(f"Loading template config from: {template_full_path}")
